+++ Tetris Monogame/GameplayManager.cs
@@ -10,7 +10,7 @@ namespace Tetris_Monogame
     {
         public bool GameStarted { get; set; }
         public bool GamePaused { get; set; }
-        public double GameSpeed { get; set; }
+        public decimal GameSpeed { get; set; }
         public int Rows { get; set; }
         public int Columns { get; set; }
